create StateSerializer, an ITextStatable implementation using the Serializer class, and wire it up to the TI83 core
